requirements.txt
Flask>=2.0.1
pandas>=1.3.0
requests>=2.25.1
xlsxwriter>=3.0.0
Werkzeug==2.0.3
tiktoken==0.8.0
azure-cosmos==4.3.0
python-dotenv==1.0.0
flask_swagger_ui==4.11.1
flask-cors==5.0.0
azure-identity==1.19.0
azure-core
urllib3